Day 13
I actually cheated on the second part of this.  I went poking around
and someone on Reddit mentioned the Chinese Remainder Theorem.  I
researched that, implemented the algorithm and finished the problem.

This was after I had spent about 8 hours trying to figure out how to
do it in a non-brute force way (which would have taken years I think).